refactor(fsdp): name the metric module after MetricBuffer
