+++ miles/backends/fsdp_utils/actor.py
@@ -17,7 +17,7 @@ from miles.utils import tracking_utils, train_metric_utils
 from miles.utils.context_utils import with_defer
 from miles.utils.distributed_utils import get_gloo_group
 from miles.utils.memory_utils import clear_memory, print_memory
-from miles.utils.metric_reduce import MetricBuffer
+from miles.utils.metric_buffer import MetricBuffer
 from miles.utils.metric_utils import compute_rollout_step
 from miles.utils.profile_utils import TrainProfiler
 from miles.utils.timer import Timer, inverse_timer, timer
